separate some filetype-specific chunks into own chunk subirs
This makes it so that filetypes (and other chunks) can pull in more
specifically what they need, rather than "pull in everything and use
some small percentage of it".

We may have to bump some of this back up a directory if it ends up
being used by more things than what we originally thought.