wowdump: implement offboard metadata & use for simplifiers (#6)
A mostly complete (if really ugly) implementation of separate
metadata that can be used for (at the moment) choosing what
simplifier is used for a data type, so that we don't have to
do a billion regex matches just to display data. Probably shaves
~15% off of dumper runtime, at least on things that have a ton
of data.

Kinda sloppy at the moment (lots of not-DRY), but seems mostly
functional and doesn't do anything *too* insane, I don't think.